abstract buildintegeratednugetproject
